Removal of unneccesary files.  ToDo: Check all comments and documentation.  Final check.  Merge and Pull request.
